customkeeps_backend/requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
stripe
sqlparse==0.5.3
tzdata==2025.2